Add option to skip running builds in Docker
By setting the envvar `NO_DOCKER` make will not run the given target
in a Docker container. This way builds are much faster because no
Docker image is built and no container has to be started.

On macOS `deadcode` complained about an unused function so that I had
to move that function into a file of its own and add build tags to it
so that it isn't compiled on macOS.

Also I removed some unused struct fields in macOS-specific code.

Fixes https://jira.mesosphere.com/browse/DCOS_OSS-5007
